Changed variable location to GameStart
