+++ src/Model/AchievementsButton.java
@@ -1,6 +1,6 @@
 package Model;
 
-import View.GameView;
+import View.GameStart;
 import javafx.scene.control.Button;
 
 public class AchievementsButton {
@@ -12,10 +12,10 @@ public class AchievementsButton {
 
 
     public AchievementsButton() {
-        buttonX = GameView.getSceneWidth()/3;
-        buttonY = GameView.getSceneHeight()*6/9;
-        buttonWidth = GameView.getSceneWidth()/3;
-        buttonY = GameView.getSceneHeight()*0.75/9;
+        buttonX = GameStart.getSceneWidth()/3;
+        buttonY = GameStart.getSceneHeight()*6/9;
+        buttonWidth = GameStart.getSceneWidth()/3;
+        buttonY = GameStart.getSceneHeight()*0.75/9;
         Button achievementsButton = new Button("Achievements Graphic needed");
         achievementsButton.setPrefSize(buttonWidth, buttonHeight);
         achievementsButton.setLayoutX(buttonX);
