+++ src/View/MainMenuView.java
@@ -1,0 +1,5 @@
+package View;
+
+public class MainMenuView {
+    
+}
